Task 1
Create an application that can replace the home launcher in your smartphone.
The project should be split into two modules:
·SDK - All query functions for all data required in the UI/App.
·App - UI implementation for listing and launching applications.

SDK

Launcher SDK will return all the launcher apps list which are currently installed on the phone. Here are the main features of the SDK:
1.List of apps containing following information - App name, Package name, Icon, Main Activity class name, Version code, and Version name.
2.The list should be in ascending order based on app-name
3.Notify when app installs/uninstalls.

UI implementation for listing and launching applications
1.List application data in a recycler view.
2.Launch the application when clicked on the app icon/list item.
3.Add a search bar on top which will filter the application list based on the name.